refactors profile editing
